--- Model/MeshExtraction.py
@@ -1,10 +1,0 @@
-import torch
-import open3d
-import mcubes
-
-import Rendering
-
-
-class MeshNeRF:
-    def __init__(self):
-        pass
