+++ Utils/MeshExtraction.py
@@ -1,0 +1,3 @@
+class MeshNeRF:
+    def __init__(self):
+        pass
